Update Dockerfile to use OpenEMR 7.0.3 and add permissions fix script
This commit updates the base image in the Dockerfile to OpenEMR 7.0.3 and introduces a new script, fix-permissions.sh, to manage permissions for the sites directory. The script is set to run at container startup, ensuring that permissions are correctly applied while maintaining security. These changes enhance the initialization process for OpenEMR by automating permissions management.