Fixes #8423, #6889 logo customizations
Adds a logo filter event to allow for customizing the web path location
of the logo.  It still requires a valid path (so if using something
dynamic from the db will need an php controller file that is publicly
accessible to serve up the dynamic image).

Updates the LogoService to dispatch the filter event.

Also adds in folders for customizing the portal primary and secondary
logos.

fixes #6889
fixes #8423

Partially addresses #7780 by allowing logo customizations but not in the
way the issue discusses.