Inferno test suite ci engine
Got the system to build and run the test on the first invocation of
run.sh.

Tearing down the dockers and re-running run.sh fails on the redis server
with the permission changes.  Redis leaves the data directories as a
999:999 uid:guid which the current user doesn't have permissions to copy
the files over and the script fails. So far don't have a solution that
fixes it outside of just removing the
onc-certification-g10-test-kit/data/redis folder after tearing down the
dockers.

It also looks like the test suites fail at times due to db contention
when the test is running and when I query for the test results in the
phpunit test suite.  I need to figure out if there is a way to run it
w/o contention.

Next steps are to figure out a ci workfile and then to also make sure
the coverage engine is turned on.