feat(auth): add gcip_enabled global setting
This commit adds a new global setting, , to the  file. This setting will be used to enable or disable the Google Cloud Identity Platform (GCIP) integration feature.

This is the second step in the GCIP integration plan, and it allows administrators to control the availability of the new authentication method.