feat(auth): add redirectToGoogle and handleGoogleCallback methods to AuthorizationController
This commit adds two new methods, redirectToGoogle and handleGoogleCallback, to the AuthorizationController. These methods will handle the Google Cloud Identity Platform (GCIP) OIDC flow.

- redirectToGoogle: This method will be responsible for constructing the authentication request to Google's OIDC endpoint and redirecting the user.
- handleGoogleCallback: This method will handle the callback from Google, exchange the authorization code for tokens, validate the ID token, and provision/authenticate the user in OpenEMR.

Additionally, the oauthAuthorizationFlow method has been modified to check the identity_provider of the client. If it's set to 'google', the user will be redirected to the redirectToGoogle method.