Get fully working metadata endpoint for unit tests.
Changed the fhir metadata endpoint to no longer api log since handled by
listener.

Moved apiType url check methods (isFhirRequest, isApi, etc..) to
HttpRestRequest method so can be used throughout the application.

Added helper method HttpRestRequest::getRequestPathWithoutSite so we can
grab the suffix path after the site.

Added an ApiResponseLoggerListener to handle all of the route logging
that is scattered throughout the routes.

Made the SkipAuthorizationStrategy populate the userId attributes and
other request instance attributes so we track the user in the request
even if they are requesting a non-authenticated request.  Values are set
to null if user is not authenticated.

Fixed up the FhirMetadataRestController to not rely on RestConfig since
we can't easily pull it in via PSR.

Fixed up the other listener classes so that we finally have a working
implementation from the dispatch.php refactoring.