Implement integration testing
ADded in an integration testing that does not run the regular
bootstrap.php which in turns includes the globals.  Since we are running
globals through the site setup listener for our api application we don't
want to use the regular phpunit process.

However, ran into an issue where the require_once makes it so I can only
have a single test that uses the SiteSetupListenerTest as the
require_once will never run again.  We need to figure out how to
abstract the globals initialization steps better so we can run the setup
process multiple times across integration tests without having to do a
full e2e setup.