Fix phpstan errors.
Number of missing return statements, variables not having definitions
and a missing constructor.

However there is still a function missing display_layout_rows_group_new
that can't be found in the codebase and I'm not sure how the the code is
even reached. I'm going to leave it in in case there is some kind of
dynamic function being defined but I can't tell how that would be.