Openemr fix phpstan 2025-07-10 (#8590)
* Fix FHIR phpstan errors

* Fix phpstan in dupecheck and mark files deprecated

* Fix phpstan issue in code types

* Fix phpstan issues in Atlas labworks export

* Fix phpstan issues in old qrda export functions

* Fix phpstan errors in telehealth module

* Fix phpstan issues with telehealth module

* Fix psr style issues

* Fix phpstan errors.

Number of missing return statements, variables not having definitions
and a missing constructor.

However there is still a function missing display_layout_rows_group_new
that can't be found in the codebase and I'm not sure how the the code is
even reached. I'm going to leave it in in case there is some kind of
dynamic function being defined but I can't tell how that would be.

* Fix phpstan issues in zend modules

Phpstan identified dead code, missing variable definitions, and missing
class imports.  Fixed all of those.

* Remove TreeMenu::createFromXML due to broken func

The createFromXML relies on a file existing in the codebase that doesn't
exist.  Did a search and the function doesn't exist so removing it.

* Cleaning up phpstan issue in prescription

* Fix Automated and Clinical Quality Measure phpstan

* Fix direct message phpstan var

* Remove unused variable in pnotes

* Fix cdr phpstan issue

* fix missing params oeHttpRequest/oeHttpResponse

* Fix missing func definition error

* Removed unused method in evidence dsi

* Fix covariant violations in fhir parser map

* Fix missing includes and variable definitions

* http rest request fix phpstan violations

* FHIR rest fix missing var definitions,export error

Export errors were not being fully tracked due to a subtle bug that
phpstan caught.

Also fixed a lot of missing variable declarations in the rest endpoints

* Removed dead code on smart fhir search endpoint

* Change ccda runtime exception to be right namespac

Namespace was incorrect for the XmlExtended class for Cda

* Fix undefined variable in cda import/dispose

* Leveraged fhir base trait and removed unused code

Fixed a bunch of phpstan issues by just using the
FhirServiceBaseEmptyTrait class in places we weren't using it and
implemented the return methods in the trait.

* fhir $doc-ref service update w/ puuid

Had an issue where the puuid wasn't being cast into the $doc-ref
operation.  Fortunately the uuid narrowed it down, but needed to enforce
patient context on the query.

* rest services, qdm,qrda phpstan fixes

Fixed phpstan errors in the rest services, qdm and qrda classes where
variables were missing, return statements were missing, and class
definitions were missing.

* Remove unused non-functioning class

The TableSearchProcessor has all kinds of broken methods and can't
possibly be used.  Removing from codebase.

* Rest api & FHIR fix phpstan issues

Fix missing test imports on rest and api unit tests.  Fix minor phpstan
issues on tests.

Fix phpstan issues in TokenSearchField and TelemetryService

* Fix phpstan issues in practitioner service test

* Removing old unit tests that aren't executed

No point in dealing with phpstan issues for old unit tests that are not
executed and have little to no utility.

* PSR fixes and phpstan fixes

* Facility service test Fix missing imports

* Fix phpstan issues on traits

* The CheckCreateUser is deprecated and redundant

We already have e2e tests for creating a user.  The e2e test here can't
even run so going to remove it.

* More phpstan fixes

* Fix more phpstan with ccda

* Fix psr... phpstorm should have caught this.

* Change ifelse to match statement

* Fix psr styles