Fix mod_rewrite changes, errors, scopes v2
The mod_rewrite changes were breaking all of the nginx configuration
settings.  I decided instead of breaking everyone's apache vhost and
nginx configurations to follow more modern practices for mod_rewrite to
keep with the existing rewrite using the query params to handle the
path.  I map the HttpRestRequest::createFromGlobals function to detect
the mod_rewrite and translate it into the Symfony and php PATH_INFO
parsing that is used typically for a rewrite command so the request
objects will have their path variables setup properly.

Also fixed up some errors in the tests and an issue with the webroot in
the SiteSetupListener.

Started the preliminary work for handling the v2 scopes from the
HttpRestRequest.  I'm thinking we'll move the logic into ScopeRepository
for better logical cohesion, but for now will keep it here.