API Scope processing changes, cleanup code
Removed unused discover.php and jwk.php files now they are handled by
actual controllers.

Implemented the ResourceScopeEntityList that holds a list of parsed
scopes for a given context/Resource so our ScopeRepository validation
array can walk through permissions, operations and subsets of a resource
constraint easily.

Fixed up the scope parsing logic in the
ScopeEntity,ScopePermissionObject, class.

I made the ServerListEntity class be the main container for what scopes
are defined in the server and the ScopeRepository is the one that
applies the logic of what scopes are valid and finalized.  This
separated out the logic and made it easier to use the server supported
scopes in other areas of the codebase without have to instantiate the
heavier ScopeRepository class.

Removed the site:{site} syntax from the scopes as the site can be pulled
from the URL and the encryption keys make sure there is no security risk
of an access token being used on the wrong site.

Moved the Claims logic out of the ScopeRepository because they are
separate in OAuth2 parlance.  We still have some superglobals going on
with the claims stuff that will have to be migrated out.

Still need to fix up the finalizeScope pieces as it doesn't handle sub
scopes yet.

Cleaned up a lot of methods in the ScopeRepository that are no longer
being used.

I changed up the logic so that ScopeRepository just deals with all of
the FHIR/standard apis and doesn't try to limit the scopes returned
based upon what API type it is.  This is a responsibility of the access
system and can fully be handled based upon standard vs fhir api as there
shouldn't be a need to separate the two as far as scope access goes.
Either the scopes exist or they don't.