fix: Use selenium/standalone-chromium to enable support for both AMD64 & ARM64
This change pins to v4 of the upstream Selenium Grid with the latest compatible version of Chromedriver. It is using Chromium as the base versus Chrome. The primary issue with the selenium/standalone-chrome image is that it does not ship an ARM64 image thereby breaking the docker compose on ARM64 like MacOS.