Scope description fixes, oauth2 claim fixes
Fixed scope permission lookup to handle both v1 and v2 permissions.

Changed up the UserEntity object to remove globals reliance.  Also
changed it so it returns claims always in the context of the smart on
fhir api usage instead of client_credentials grant as its never called
in that code pathway anyways.

Refactored the UserRepository so it doesn't extend from UserEntity which
was a bit strange.  Made UserRepository responsible for grabbing the
UserEntity class with the getAccountByPassword function.

Changed up IdSmartTokenResponse to add in the nonce claim since its more
of a system-wide scope that doesn't rely on UserEntity for the claim.

Added ClientRepository::remove so we can clean up clients for testing
purposes.