Upgrade admin-client page, update token repo
Updated the token repo to use the SessionStorage interface so we can
unit test with it.

Changed up the admin-client page to return Response objects so we can
actually test the implemenation instead of exiting and dieing throughout
the codebase.

Fixed up a bunch of the broken unit tests.

Removed unused code in the RestConfig class.

Used claude to write an extensive test for the ClientAdminControllerTest class.