More scope v1/v2 access fixes on api
The v1 compatability layer on ScopePermissionObject was not working
properly once hooked up to the scope check on the specific api access
with the v2 scopes (cruds syntax vs read/write).

Changed up the ScopePermissionObject to handle multi-operation scopes so
the scopeContains operation in ScopeEntity works properly.

Fixed the SkipAuthorization and LocalApi authorization pieces as well as
requests were failing against metadata / version, product endpoints.

Populated the resource Instance Identifier in the HttpRestParsedRoute so
we can differentiate on a FHIR 'read' operation vs a FHIR 'search'
operation. Kept the same methodology for the standard api.

Fixed a number of phpstan issues in the HttpRestRequest class as well as
in some consumers of the class.

Marked HttpRestRequest::setAccessTokenScopes as deprecated so we can
work with ScopeEntity objects and not strings.

Set a _route property in the HttpRestRequest->attributes that will have
the valid parsed HttpRestParsedRoute with all of its attributes.

Set HttpRestRequest to use the scope validator factory method instead of
handling the logic manually inside the class.