WIP SDOH fhir observations
Implemented two new mapped observation classes.  One for the Observation
Form and one for the Questionnaire Response items.  Right now the
parsing of the observations is all complete but we need to build out the
underlying data models in OpenEMR to accomodate the changes.

I also implemented a unit test for the CodeTypesService class so we can
have better test coverage there.