Observation saving, questionnaire linking
Got the observations saving and questionnaires linking properly.  Fixed
up navigation flow between new, edit, and list.  Fixed up the reports to
show more of the actual observation details.

Did prep work for sub-observations which is my next task list.

Commented out a number of search and filtering functionality that can be
added in later, but for now, just trying to focus on the core
requirements needed for US Core.

Fixed an issue with the fhir questionnaire endpoints making them not
work properly.

Implemented the questoinnaire search and linking for an observation in
order to handle the observation justification.