Fix observation date formatting.
The date formatting was not being handled properly for observations.  We
were missing the javascript libraries needed and the formatting was not
saved properly.  This wasn't caught when the default Y-m-d H:i:s date
string was used.  But when using a different format everything breaks.

Had to updte the DateFormatterUtils in order to support the time seconds
piece.