feat(ServiceLocator): TelemetryService Initial
This change introduces a service locator approach to dependency injection. Instead of directly instantiating the class, we use the service locator to find registered implementations of the interface. The default implementation is the original class already in OpenEMR, as registered in `config/services.php`. Then we provide an event so that custom modules can inject alternative implementations, without needing to modify the register directly.

For the initial implementation we use the well-tested TelemetryService, which should let us know if this is fundamentally working.