Capability Statement - Profile Versioning
To pass the capability statement for observations, I needed to overhaul
the profile versioning.  I went ahead and added it to all of the
resources to support the versioning profile.  Will need to go and add
the individual meta statements afterwards, but this makes the capability
statement pass for 3.1.1, 7.0.0 and 8.0.0 US Core.