Observations for SDOH class
Implemented uuid mapping for the form_history_sdoh form that covers the
sdoh elements and the pregnancy status and pregnancy intent elements.

Refactored history save to use the HistorySdohService to consolidate the
code and allow the uuid mapping to function properly.

Fixed some broken update scripts in the 7.0.3 upgrade.
Added the pregnancy_intent list type.

Added category field to the questionnaire_repository.

Fixed up a bunch of integration tests that were broken with the
disabling of the QuestionnaireResponse endpoint for observations with
the inferno validation problems.