FHIR Fix sdoh mapped codes
The SDOH mapped codes were not returning the correct resource when
searching for a specific mapped instance (getOne).  This was because the
codes were not being constrained in the search inside the category but
were treated as a UNION and the wrong resource was then grabbed as the
first result found.