Fix Occupation FHIR Observation
In order to handle the start and end date of employment had to refactor
how employer_data is handled.  I changed up the occupation and industry
so they get copied in the employer_data and so the table now represents
a full historical record of the patient's employment history.  I moved a
number of the employer_data methods into the EmployerService class.