WIP average blood pressure
Implement the foundational framework for average blood pressure by
allowing for calculated vitals records that are computed as a derivative
to primary source vital calculations.

Right now this calculates 3day mean, last 5 vital readings, and mean of
all vitals in an encounter.

Still need to handle the aggregation of bps and bpd which is a weakness
with this current approach.  I may need to break the tables up into a
calculation and then have sub columns that are part of that calculation.