feat: SDOH fhir observations #8792 (#8837)
* Refactor observation files into MVC architecture

Implemented a Controller for the new,report,save views of the
observation form.  Implemented a service class to put observation
retrieval and saving into.

Wrote unit tests to ensure the basic functionality works before we add
additional data elements to the observations.

* Fixing code review issues, phpstan issues

Added AI generation comments, fixed phpstan complaints.

* Remove unused tests, refactor queries

Refactored the queries to use QueryUtils and changed tests that had
little utility in order to mark them incomplete for the future.

* Observation top level deletion

Got the top level deletion working for observation.  In reality we set
the activity flag to 0 on the observation instead of whole-scale
deleting them.

I now need to handle the sub-observation deleting which will be the
intersection between the new and old on the postData.

* Fix unit tests, phpstan, styles

* WIP SDOH fhir observations

Implemented two new mapped observation classes.  One for the Observation
Form and one for the Questionnaire Response items.  Right now the
parsing of the observations is all complete but we need to build out the
underlying data models in OpenEMR to accomodate the changes.

I also implemented a unit test for the CodeTypesService class so we can
have better test coverage there.

* Observation Implement us core resource hydration

Implemented the US core parseOpenEMRRecord while handling the
derivedFrom and memberOf attributes.  Had to rework the search
capabilities of the observation form to make that possible but now
category and fhir search endpoints should be working.  Just need to test
it against the inferno test suite now for observations.

* Adding api test and access token commands

Adding in helper commands so I can continue the FHIR testing for
Observations

* Fixed no data found exception in fhir

Found an error when searching sub-observations if the parent observation
has no id.

* Fix observation date formatting.

The date formatting was not being handled properly for observations.  We
were missing the javascript libraries needed and the formatting was not
saved properly.  This wasn't caught when the default Y-m-d H:i:s date
string was used.  But when using a different format everything breaks.

Had to updte the DateFormatterUtils in order to support the time seconds
piece.

* Capability Statement - Profile Versioning

To pass the capability statement for observations, I needed to overhaul
the profile versioning.  I went ahead and added it to all of the
resources to support the versioning profile.  Will need to go and add
the individual meta statements afterwards, but this makes the capability
statement pass for 3.1.1, 7.0.0 and 8.0.0 US Core.

* Fix vitals, observation profile versions

Fixed issues on vitals w/ US Core 7.0.0 & 8.0.0.

Added observation profile versions.

* Fix PSR issues

* Access-Token generator, fix command-runner

Access Token command was broken with the code from master route
refactor.  Also made it so the symfony commander runner will inject the
globals.

* Fix profile issue on head circumference

* SDOH Observation implementation.

Fixed questionnaireresponse to handle single item.
Fixed observation form to handle status saves
Added db fields for observation to allow values to be a codeset value.

Handled 404's with Observations better.

Got the observation screening-assessment with us core fully functionning
which was a pain.

Implemented observation performers.  Fixed a number of code issues with
the observation form.

Running into an issue with the QuestionnaireResponse.  It appears the
system fails when it attempts to grab the Questionnaire resource and
throws a fatal inferno error.  Not sure what is going on but the tests
pass w/o QuestionnaireResponse so perhaps its an inferno thing.

* Rector and unit test fixes

* Disabling observation questionnaireresponse

Disabling the observation questionnaire response until we can figure out
the FHIR validation errors being thrown by inferno.  Oddly inferno still
passes with it disabled.  I think its because the Questionnaire resource
is not publicly accessible but I don't know for now.

* Add encounter reference to vitals

* Fix register client command

* Observations for SDOH class

Implemented uuid mapping for the form_history_sdoh form that covers the
sdoh elements and the pregnancy status and pregnancy intent elements.

Refactored history save to use the HistorySdohService to consolidate the
code and allow the uuid mapping to function properly.

Fixed some broken update scripts in the 7.0.3 upgrade.
Added the pregnancy_intent list type.

Added category field to the questionnaire_repository.

Fixed up a bunch of integration tests that were broken with the
disabling of the QuestionnaireResponse endpoint for observations with
the inferno validation problems.

* Fix uuid,date searches, category+code searches

Had issues with the category+code searches

Vitals uuid was failing to encode properly when a non-matching resource
was present.

Fixed category and codes not working together in social history and sdoh
service.

* Fix codeable concept for pregnancy intent+status

* More fixes on smoking status history

Had to fix up code representations on smoking status.  Looks like there
is still one missing element with the valueQuantity which we need to be
able to represent the years someone has smoked or using a nicotine
product.

* FHIR Fix sdoh mapped codes

The SDOH mapped codes were not returning the correct resource when
searching for a specific mapped instance (getOne).  This was because the
codes were not being constrained in the search inside the category but
were treated as a UNION and the wrong resource was then grabbed as the
first result found.

* Duplicate survey category code in observations

Fixed it so the survey category code is not being sent twice.

* Fix code mappings, phpstan issues

Got the screening assessment piece working properly.  I'd still like to
come back and map QuestionnaireResponse.items to an actual table and
make it searchable.  That would make everything much more usable, but
the SDOH assessment handles things and the regular form_observation
while cumbersome can meet the task for now.

* Fix US Core 3.1.1 for smoking status

Issued was a mandatory field for 3.1.1 so we need to add it back in to
pass certification.

* Fix internationalization formatting

* Fix types and redunant code from rector

* Fix sql errors and rector issues

* Update trait order for style check

* More style fixes, fix bad variable name

* Fix escaping, extraneous comments

* Fix rector issues with array() syntax

* Fix rector strpos issues...