Feat openemr 9035 9036 patient observations (#9039)
* Fixes #9035, #9036 patient observations

Fixes #9035 - Implemented the Sexual Orientation Observation
Fixes #9036 - Implemented the Occupation Observation.

Reworked the FhirObservationTrait to handle component pieces for the
Occupation.  Could potentially rework the vitals handling to leverage it
at some point.

* Fix Occupation FHIR Observation

In order to handle the start and end date of employment had to refactor
how employer_data is handled.  I changed up the occupation and industry
so they get copied in the employer_data and so the table now represents
a full historical record of the patient's employment history.  I moved a
number of the employer_data methods into the EmployerService class.

* Finish up occupation observation.

Was missing the mapped resource in main observation service.  Also date
was being double set on insert.

Fixed up the demographics save not actually saving the employment data

* Occupation period dates

Fixed the period dates for the occupation with start and end dates for
the observation period.

* Fix rector issues and effectivePeriod problem

* Fix capability statement tests

* Fix rector

* Fix employment columns in new database

New database wasn't matching the upgrade sql script for employment
columns.  Added in the layout_options.  Also Fixed the industry to match
the text value insead of longtext.  Removed the default null values.

* Fix unit tests with date value updates.

Unit test was using effectiveDatePeriod when we needed to test for
effectiveDateTime due to an update in FhirObservationTrait