WIP FHIR Condition Encounter Diagnosis
This commit implements the baseline functionality needed for Conditions
with a type of encounter-diagnosis.  In 3.1.1 we were incorrectly
labeling them as problem-items even if they had an attached encounter
diagnosis.

I had to overhaul the db structure of the conditions as the 3.1.1
implementation would create duplicate condition records if a
medical_problem lists record was attached to multiple encounters via the
issue_encounter join table.  In order to represent the 'instance' nature
of the condition linked to an encounter per FHIR spec I had to add a
uuid to the issue_encounter join relationship.  To be backwards
compatible I added a cross-over date check on the condition so that
after our deployment on November 15th conditions will use the uuid from
the issue_encounter table and fallback to the old uuid for anything
before that date.

In order to accomplish this I had to centralize all the insert queries
everywhere issue_encounter is created (fee sheet, eye form, patient
issues). This allowed me to generate the uuids properly.

Our UUIDRegistry does not handle composite primary keys so I had to add
a surrogate primary key to the issue_encounters table that
auto_increments in order to get the UUIDRegistry to work properly
without overhaulings its functionality.  If a system has a large number
of linked encounters to issues this may increase the runtime of the
sql_upgrade.php.

I added in the FhirConditionProblemsHealthConcernService as well but
need to do further testing there before I'll consider that class
complete.