Unit test fixes, provenance info.
Was missing a few columns in 8.0.0 implementation in order to handle
provenance.  Added those in and updated PatientIssuesService everywhere
the session was being used.

Fixed up the original FhirConditionEncounterDiagnosisServiceTest class.
probably need to merge it into the 8_0_0 test but that one is more of an
integration test whereas the FhirConditionEncounterDiagnosisServiceTest
is a unit test on the parsing.