Openemr fix 8787 fhir condition encounter diagnosis (#9117)
* US Core 8.0.0 fhir Condition resource.

Fixes #8787.

Changed up the FhirCondition logic to use the Mapped resource pattern
like we do with the Observation resources.  The condition is split up
into the encounter diganosis and problems/health concerns service
classes.  We may need to split problems/health concerns up even more
depending on where we pull the logic from for these resources.

Created a new Trait MappedServiceCategoryTrait for resources that only
use categories and not codes.

Implemented a ProfileMapper that I originally thought I was going to use
for adding / manipulating the data in the profile, but so far I haven't
had to do any of that so we may just remove the entire class.

Also wrote unit tests for the sub-resource classes for Condition.  Still
need more unit tests to handle additional code coverage.

* Condition tests to make sure 3.1.1 doesn't break

* Left out unit test...

* WIP FHIR Condition Encounter Diagnosis

This commit implements the baseline functionality needed for Conditions
with a type of encounter-diagnosis.  In 3.1.1 we were incorrectly
labeling them as problem-items even if they had an attached encounter
diagnosis.

I had to overhaul the db structure of the conditions as the 3.1.1
implementation would create duplicate condition records if a
medical_problem lists record was attached to multiple encounters via the
issue_encounter join table.  In order to represent the 'instance' nature
of the condition linked to an encounter per FHIR spec I had to add a
uuid to the issue_encounter join relationship.  To be backwards
compatible I added a cross-over date check on the condition so that
after our deployment on November 15th conditions will use the uuid from
the issue_encounter table and fallback to the old uuid for anything
before that date.

In order to accomplish this I had to centralize all the insert queries
everywhere issue_encounter is created (fee sheet, eye form, patient
issues). This allowed me to generate the uuids properly.

Our UUIDRegistry does not handle composite primary keys so I had to add
a surrogate primary key to the issue_encounters table that
auto_increments in order to get the UUIDRegistry to work properly
without overhaulings its functionality.  If a system has a large number
of linked encounters to issues this may increase the runtime of the
sql_upgrade.php.

I added in the FhirConditionProblemsHealthConcernService as well but
need to do further testing there before I'll consider that class
complete.

* Fixed up db columns added provenance info

Added the provenance columns for tracking the recorder user and author
info for conditions.

Fixed up code review suggestions of missing unique keys, and other
fields between new db and upgrade script.

* Unit test fixes, provenance info.

Was missing a few columns in 8.0.0 implementation in order to handle
provenance.  Added those in and updated PatientIssuesService everywhere
the session was being used.

Fixed up the original FhirConditionEncounterDiagnosisServiceTest class.
probably need to merge it into the 8_0_0 test but that one is more of an
integration test whereas the FhirConditionEncounterDiagnosisServiceTest
is a unit test on the parsing.

* Condition Encounter Diagnosis

This should bring everything into full implementation on the encounter
diagnosis.  Still working on the problem item and sdoh health concern
pieces.

* Fix upgrade script

* Condition searching on patient,category

Fixed issues where patient search was breaking and category if provided
broke condition search.

Also fixed broken rest controller.

* Disable Condition Health Problems for now

turning off the health problems until we can get the encounter diagnosis
passing completely.

* Condition Fix getOne call, code cleanup

Removed unused code and the experimental profile mapping stuff.  Fixed
issues with the getOne call for single resource _id fetch.

* Fix styling and rector issues

* Suppress extraneous log messages in unit tests

* Fix unit tests and clinical status date parsing

* Fix styling issues

* Fix style, rector, and phpstan issues

* Fix rector problem with last fix