MedicationDispense endpoint implementation
Changed up Prescription drug interface UX to have a toggle button for
showing the dispense attributes when the pharmacy inventory global is
enabled.

Added to dispensing the ability to set the supply type of the dispensing
(IE full first order, partial order, refill, etc).

Fixed up a number of broken behaviors that was going on dealing with the
encounter linkage for dispensing drugs.  Incorporated the changes for
encounter validation and message reporting from PR #8598 since my
changes had conflicts with that one.

Fixes #6605 dispensing medication broken
Fixes #8597 with contributions from @cburnicki  (Christoph Burnicki)

Moves the code for drug_sales into the dedicated DrugSalesService class.
Removes the drug_sales code that would just die and instead implemented
exception handling (would prefer more type specific exceptions but went
with generics at this point).

Changed up the Encounters summary screen so that if pharmacy inventory
management is enabled it will show as part of the encounter summaries
the Dispensed medication that was linked up to the encounter.

Added to the Prescriptions screen the encounter link so you can tie
prescriptions to an specific encounter (db had linkage already so don't
know if this behavior originally existed).

Major change to Prescriptions to not default the encounter id always to
the current session and instead use what is stored in the database for
existing records.

Updated the drug_route codes to have NCI codes so they come through
properly in FHIR.

Added a new TwigExtension function called encounterSelectList for
getting a dropdown of most recent encounters.

Fixed up the MedicationDispense endpoint to be compliant with US Core
8.0 requirements.
Added a new 'context' search parameter on MedicationDispense so we could
use it internally as well as provide users the ability to filter
medications to a specific encounter.