Fix NDC code system, add organization performer
We add the primary organization as a performer in the MedicationDispense
in addition to the practitioner who prescribed the medicine.

Also the display text for the codes was using the prescription name
instead of the code system text.  Changed that.  Now if the ndc or
rxnorm are not installed at least it will blank out instead of using the
less accurate drug name