Codesystems fix, timing resource fix
Timing also had the resourceType field. I'm going to guess some of the
backbone elements generated with php-carbon have a bug in populating this
resourceType field.

Changed up the codesystems since the code description can't be null,
fell back to the drug name and if that is null, then the code is
stripped.