Diagnosis reason for FHIR, prescription dups
Finally implemented a direct linkage from prescriptions to
lists_medication in order to mitigate duplicates showing up in the FHIR
api for MedicationRequest.

Implemented diagnosis in the prescriptions table in order to be able to
track the prescription reasonCode.  Made the values populate into the
lists_medication table so we can have better tracking and
lists_medication becomes the preliminary of a medication administration
record.

I kept with the existing Prescription class despite its problems but
heavily refactored the db inserts in order to get away from the
wierdness of class setter methods actually persisting data to the
database in a wierd kind of ruby active record monster.