Feat openemr fix #9183 medication dispense (#9192)
* WIP initial MedicationDispense implementation

Using the drug_sales service table as an initial implementation for the
medication dispense resource.  Wrote out a US Core FHIR compliance piece
with some AI help.

Added uuid to drug_sales.

* MedicationDispense endpoint implementation

Changed up Prescription drug interface UX to have a toggle button for
showing the dispense attributes when the pharmacy inventory global is
enabled.

Added to dispensing the ability to set the supply type of the dispensing
(IE full first order, partial order, refill, etc).

Fixed up a number of broken behaviors that was going on dealing with the
encounter linkage for dispensing drugs.  Incorporated the changes for
encounter validation and message reporting from PR #8598 since my
changes had conflicts with that one.

Fixes #6605 dispensing medication broken
Fixes #8597 with contributions from @cburnicki  (Christoph Burnicki)

Moves the code for drug_sales into the dedicated DrugSalesService class.
Removes the drug_sales code that would just die and instead implemented
exception handling (would prefer more type specific exceptions but went
with generics at this point).

Changed up the Encounters summary screen so that if pharmacy inventory
management is enabled it will show as part of the encounter summaries
the Dispensed medication that was linked up to the encounter.

Added to the Prescriptions screen the encounter link so you can tie
prescriptions to an specific encounter (db had linkage already so don't
know if this behavior originally existed).

Major change to Prescriptions to not default the encounter id always to
the current session and instead use what is stored in the database for
existing records.

Updated the drug_route codes to have NCI codes so they come through
properly in FHIR.

Added a new TwigExtension function called encounterSelectList for
getting a dropdown of most recent encounters.

Fixed up the MedicationDispense endpoint to be compliant with US Core
8.0 requirements.
Added a new 'context' search parameter on MedicationDispense so we could
use it internally as well as provide users the ability to filter
medications to a specific encounter.

* Fix json serialization with resourceType

* Fix new install issue with drug_route

* Globals broken due to incorrect enum

Was using const instead of case on enums.

* Was missing type reference

* Fix getOne search

Uuid was not unique, had to make the columns unique to search on them.

* Add provenance and performer pieces

* Update drug_interval with codes and titles

* Add timing, drugAndDose to MedicationDispense

Added the timing informatin, with drugAndDose's quantity information.

* Handle drug timing with no valid FHIR code

Set the interval_notes from the drug_interval list if there are no valid
mapped fhir codes.

* More drug_interval db work

Fix it for new db instances, update codes / descriptions for existing
drug_intervals

* Fix quantity problem

* Fix MedicationDispense.quantity.unit

The unit was using the dosage unit from the prescription instead of the
drug unit which in this case would be tablet (or package, or whatever
the container is).

* Fix NDC code system, add organization performer

We add the primary organization as a performer in the MedicationDispense
in addition to the practitioner who prescribed the medicine.

Also the display text for the codes was using the prescription name
instead of the code system text.  Changed that.  Now if the ndc or
rxnorm are not installed at least it will blank out instead of using the
less accurate drug name

* Codesystems fix, timing resource fix

Timing also had the resourceType field. I'm going to guess some of the
backbone elements generated with php-carbon have a bug in populating this
resourceType field.

Changed up the codesystems since the code description can't be null,
fell back to the drug name and if that is null, then the code is
stripped.

* Fix unit tests

* Fix styling bugs

* Fix rector issues

* Fix phpstan issues

* Remove redundant exception

* Fix namespace usage

* Fix escaping issues

* Fix missing db columns, update logic

The update logic wasn't handling if a user updates the titles for the
medication timing.

Also was missing column entries in fresh db install