Feat openemr uscore8 fix procedure (#9241)
* Fix meta profile urls

Profile urls for procedures were either missing or null values.  Fixed
the issues

* Fix performed[x] and status

Fixes the datePerformed field for the different fhir profile versions.
3.1.1 requires a Data Absent Reasons, but later profiles just skip the
field.

If there are no valid dates, the status cannot be pending or completed
as that requires a performedDate.  We set the status to be unknown if
there is a pending/completed status but no valid date as we really can't
tell what the status is at that point as we have bad data.

* Fix date_ordered using wrong field

date_ordered was wrong so performedDate was coming up as current
timestamp

* Fix report date searching