Feat openemr fix #9159 medication request (#9237)
* WIP #9159 Medication Request implementation

Added Medication Adherence section to medication lists screen. Users can
now report on the reporting source of the adherence, compliance status,
and date the adherence was reported.

Had to create OpenEMR models for DateTime to handle the invalid
resourceType property.

Heavily refactored the MedicationRequest to comply with US core
requirements.  Currently have stubbed out the reason and reasonReference
until we can fully implement that requirement.

* US Core MedicationRequest unit tests

Initial AI file generated before I start doing heavy modifications.

* Fix invalid service reference

* Fix medication list dropdown

Fix label and dropdown list not showing for med adherence

* Handled case where intent was null.

* Inferno medication adherence fix

The medication adherence in inferno doesn't like Data Absent Reasons if
sub-components are missing.  So I skip the entire extension if its not
there.

* Fix dosageInstructions.

The MedicationRequest dosage instructions was not working properly

* Fixed medication adherence issues

* Fix dosage instruction quantity

* Add quantity for dispenseRequest

* Fix adherency fhir system bug

System wasn't being populated due to use of SNOMEDCT instead of
SNOMED-CT for code system.

* Fix mismatched columns on quantity

* Fix date issues with adherence

* Fixing numeric quantity on dispenserequest

* Adding in unit test class

* Diagnosis reason for FHIR, prescription dups

Finally implemented a direct linkage from prescriptions to
lists_medication in order to mitigate duplicates showing up in the FHIR
api for MedicationRequest.

Implemented diagnosis in the prescriptions table in order to be able to
track the prescription reasonCode.  Made the values populate into the
lists_medication table so we can have better tracking and
lists_medication becomes the preliminary of a medication administration
record.

I kept with the existing Prescription class despite its problems but
heavily refactored the db inserts in order to get away from the
wierdness of class setter methods actually persisting data to the
database in a wierd kind of ruby active record monster.

* Fixing empty usage/intent columns

* Update unit tests

* Fix lists_medication insert statement

Fix the insert statement so lists_medication is created properly

* Remove diagnosis codes as drug codes for meds

The lists.diagnosis was being used as the medication drug code which is
incorrect as lists.diagnosis are for indications / reasons and NOT for
drug codes.  We use the title of the medication as the text description
of the drug.

* Remove duplicates in codes selector list

The codes selector list in the prescriptions gui was showing duplicate
diagnosis if the lists table had multiple issues that have the same
diagnosis.  Had to dedup the list.

* Fix styling and phpstan issues

Ran phpcbf on src & tests... overlooked library apparently

* Fix phpcbf issues