WIP DiagnosticReport US Core 8.0 #9250
This is a preliminary WIP for DiagnosticReport.  We have to handle
linking to the observation results for the clinical notes as well as the
actual image files that the diagnostic report documentation is about.

Right now I'm just stubbing the data out for the linkage but we'll need
to add it to the core services / db / and the UX to establish the
linkage.

I added a new /fhir/Media endpoint in order to handle the DICOM and
image files for a patient.

I also discovered a bug where the DocumentService was using the globals
session variable in the REST apis (which is not populated) instead of
the SessionStorage pieces.  Fixed that area up and added a setSession
method to the BaseService and FhirBaseService so we can use the session
objects.

Fixes #9250