DB fixes, phpstan/rector UI fixes
Number of fixes to get the database and the UX to work properly with the
related persons.  Still don't have address / telecom saving for related
persons so will need to address that tomorrow.

We also have outstanding data migration from jerry's table into the
person and contact forms.

Also need to make sure all the tables are in the database.sql for new
databases.  I believe outstanding is still the layout form insertions.