CareTeam refactor bug fixes, contact extension
This builds on a lot of work that Jerry Padgett has done.

I Refactored care_teams table into two tables, care_team and
care_team_member.  care_team tracks the status of the care team, notes,
and the team name.  care_team_member tracks the individual members of
the care team. This solves a pretty nasty bug if someone removes the
first care team member from the care team and the uuid of the entire
care team ends up being changed which is not something we want from a
FHIR resource usage.  It also facilitates the status of the entire care
team.

Consolidated a number of the queries for the CareTeam view and usage.
Fixed a bug where care teams weren't handling removing all the members
of a care team.