CareTeam overall status, bug fixes
Added the ability to set the status of the CareTeam.  I made the
entered-in-error care teams not show up on the visual display where
everything else gets displayed.

Fixed bugs on save, removed the key off of the team name and instead
used the team id which paves the way for having multiple care teams in
the future (such as a proposed, active, suspended, etc).