Feat openemr 8789 condition (#9307)
* Rename condition in prep for implementation

* Fixes #8789 Health Concern Conditions

Adds a new issue type for Health Concerns to the issues list.  In the
add_edit_issue.php screen it will popup a different sub type list for
the Classification so that health concern categories can be chosen.

Fully implements the Condition piece and passes inferno testing for
/fhir/Condition endpoints with category=problem-list-item or
category=encounter-diagnosis or category=health-concern.

Adds a new feature in the SDOH assessments that will present a list of
health concern issues to add to the patient's issue list if an
assessment reaches the assessment 'positive' threshold.  This matches
what is in the ccda encounter screen.  It does not take the CarePlan
health concern and do the same thing but the NationNotes feature in
CarePlan health concern allows the notes.related_to to be linked to a
Health Concern issue to provide a linkage there if people want to link
the care plan health concern with the issues health concerns identified
there.

Fixes up incompatibilities in the code system usage between the 3.1.1
code system used for problems and health-concerns and the 3.1.1+ (7 and
8 of us core).

* Fix escaping, rector issues

* Fix return url when no concerns selected