Fixes #8641 OAuth2 asymmetric authentication (#9374)
* Fixes #8641 OAuth2 asymmetric authentication

Fixes #8641.  authorization_grant, refresh_grant, and token inspection
all now support authentication via asymmetric authentication (either
with a registered JWKS - not recommended or via a JWKS Hosted URL).

Took all of the JWK Validation code out of CustomClientCredentialsGrant
and put it in the JWTClientAuthenticationService class.

* Fix up test suites

* Fix rector

* More rector fixes

Wierd it didn't detect this one before

* Make group export follow same env as other api

group export failing on github dockers.  I believe this is due to the
URL so giving this a go.