Fixes #9409 handling practitioner for reported[x]
Fixes #9409 by have an option to say the record is not a primary record and
to choose a user from the address book. At some point when we disentangle
external organizations from the address book and put them in facility we can
handle external organizations in addition to practitioners. If a user is
chosen without an NPI it will show the Primary Business Entity as the
reported organization.