Fixes #9368, #9423
Fixes #9368 patient context population for patient context restriction
Fixes #9423 fix token serialization / deserialization

The token was not deserializing in inferno due to the + character from
encryption so needed to base64 encode it.

Also left out the context population from the rest api refactor. Made it
so the patient context gets populated on the request